Added word_list check to is_valid_word and updated test suite
Tests for play_hand indicated that the is_valid_word function was only testing for whether the word matched the hand, and not whether the word itself was a valid English word. I added this functionality into the is_valid_word function and added a test that inputs a misspelled word into is_valid_word and checks whether it comes back false.